Extracted `HoodieInputFormatBase` holding common functionality of listing FS for to be shared across File-format specific implementations   - Snapshot queries   - Incremental queries
